dev-requirements.txt
ckantoolkit==0.0.4
pytest-ckan
pytest-cov
mock
